implemented subDimensionCombinator

+++ src/parametersControl.cpp
@@ -898,7 +898,7 @@ void parametersControl::loadPreset(string presetName, string bank){
         }
         
         //if we have to destroy the module we do it
-        if(hasToBeDestroyed && moduleName != "senderManager" && moduleName != "waveScope" && moduleName != "audioControls" && moduleName != "colorApplier" && moduleName != "chartresTextureUnifier" && moduleName != "oscillatorGroup" && moduleName != "dataRecorder" && moduleName != "audioAnalyzer" && moduleName != "oscInput" && moduleName != "speakerPowerCalculator" && moduleName != "textureUnifier"){
+        if(hasToBeDestroyed && moduleName != "senderManager" && moduleName != "waveScope" && moduleName != "audioControls" && moduleName != "colorApplier" && moduleName != "chartresTextureUnifier" && moduleName != "oscillatorGroup" && moduleName != "dataRecorder" && moduleName != "audioAnalyzer" && moduleName != "oscInput" && moduleName != "speakerPowerCalculator" && moduleName != "textureUnifier" && moduleName != "subDimensionCombinator" && moduleName != "bankDimensionCombinator"){
             destroyModuleAndConnections(i);
         }
         else{
